+++ src/types/form.ts
@@ -1,5 +1,5 @@
 interface ObjectValue { value: string; label: string }
-type Value = string | ObjectValue;
+type Value<T extends string> = T | ObjectValue;
 
 export type {
     ObjectValue,