+++ src/utils/toObjectValue.ts
@@ -1,6 +1,6 @@
 import type { ObjectValue, Value } from "../types/form";
 
-const toObjectValue = (s: Value): ObjectValue => {
+const toObjectValue = <T extends string>(s: Value<T>): ObjectValue => {
     return typeof s === "string" ? { value: s, label: s } : s;
 };