+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.15.4] - 2022-04-19
+### Added
+- `className` support for `StickyHeader` and `Choice`
+### Changed
+- `Choice` is no longer all wide by default (available via variant)
+- background clipping for header, improves custom colors support
+- callback types improved for Choice
+
 ## [0.15.3] - 2022-04-07
 ### Fixed
 - `Table` `wide` variant